+++ lib/pages/settings/dashboards/dashboard_definition_page.dart
@@ -3,7 +3,6 @@ import 'dart:convert';
 import 'package:adaptive_dialog/adaptive_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_datetime_picker/flutter_datetime_picker.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/classes/entity_definitions.dart';
@@ -361,19 +360,7 @@ class _DashboardDefinitionPageState extends State<DashboardDefinitionPage> {
                                         localizations.dashboardReviewTimeLabel,
                                     labelStyle: labelStyle(),
                                   ),
-                                  theme: DatePickerTheme(
-                                    headerColor: colorConfig().headerBgColor,
-                                    backgroundColor: colorConfig().bodyBgColor,
-                                    itemStyle: const TextStyle(
-                                      color: Colors.white,
-                                      fontWeight: FontWeight.bold,
-                                      fontSize: 18,
-                                    ),
-                                    doneStyle: const TextStyle(
-                                      color: Colors.white,
-                                      fontSize: 16,
-                                    ),
-                                  ),
+                                  theme: datePickerTheme(),
                                 ),
                               ],
                             ),