+++ test/themes/themes_service_test.dart
@@ -175,6 +175,7 @@ void main() {
         'headerFontColor',
         'ice',
         'inactiveAudioControl',
+        'iron',
         'negspace',
         'outboxErrorColor',
         'outboxPendingColor',